+++ webpack.config.dev.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const webpack = require('webpack');
-const baseConfig = require("./base.config");
+const baseConfig = require("./webpack.config.base");
 const merge = require("webpack-merge");
 
 module.exports = merge(baseConfig,
